dom/indexedDB/test/test_persistenceType.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const name = window.location.pathname;
      const version = 1;

      const objectStoreName = "Foo";
      const data = { key: 1, value: "bar" };

      try {
        indexedDB.open(name, { version: version, storage: "unknown" });
        ok(false, "Should have thrown!");
      }
      catch (e) {
        ok(e instanceof TypeError, "Got TypeError.");
        is(e.name, "TypeError", "Good error name.");
      }

      let request = indexedDB.open(name, { version: version,
                                           storage: "persistent" });
      request.onerror = grabEventAndContinueHandler;
      request.onupgradeneeded = unexpectedSuccessHandler;
      request.onsuccess = unexpectedSuccessHandler;
      let event = yield undefined;

      is(event.type, "error", "Got error event");
      is(event.target, request, "Got correct target");
      is(event.target.error.name, "UnknownError", "Got correct error name");
      event.preventDefault();

      request = indexedDB.open(name, { version: version });
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let objectStore = db.createObjectStore(objectStoreName, { });

      event = yield undefined;

      is(event.type, "success", "Got correct event type");

      is(db.name, name, "Correct name");
      is(db.version, version, "Correct version");
      is(db.storage, "persistent", "Correct persistence type");

      objectStore = db.transaction([objectStoreName], "readwrite")
                      .objectStore(objectStoreName);

      request = objectStore.get(data.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.target.result, null, "Got no data");

      request = objectStore.add(data.value, data.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.target.result, data.key, "Got correct key");

      request = indexedDB.open(name, { version: version,
                                       storage: "temporary" });
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.type, "upgradeneeded", "Got correct event type");

      db = event.target.result;
      db.onerror = errorHandler;

      objectStore = db.createObjectStore(objectStoreName, { });

      event = yield undefined;

      is(event.type, "success", "Got correct event type");

      is(db.name, name, "Correct name");
      is(db.version, version, "Correct version");
      is(db.storage, "temporary", "Correct persistence type");

      objectStore = db.transaction([objectStoreName])
                      .objectStore(objectStoreName);

      request = objectStore.get(data.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.target.result, null, "Got no data");

      finishTest();
      yield undefined;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest(true);"></body>

</html>
